WorkXGoAm/flask_server/requirements.txt
flask
flask-cors
pygame
pynput
pywin32
psutil
Pillow>=10.0.0

